pyproject.toml
[tool.poetry]
name = "nebus-test"
version = "0.1.0"
description = "REST API for Nebus"
authors = ["kioharu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
alembic = "1.14.1"
fastapi = { version = "0.115.7", extras = ["standard"] }
gunicorn = { extras = ["gevent"], version = "23.0.0" }
pydantic-settings = "2.5.2"
sqlalchemy = "2.0.37"
asyncpg = "0.28.0"
geopy = "2.3.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
atomic = true
profile = "black"
py_version = 311
line_length = 88
skip_gitignore = true
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
